Addressing feedback, title and subtitle now optional, moved appropriate files into VictoryTheme folder, better importing
