+++ packages/component-library/stories/BarChart.story.js
@@ -16,30 +16,35 @@ const description = `
   This is some basic usage with the button with providing a label to show the text.
   Clicking should trigger an action.`;
 
-export default () => storiesOf(displayName, module).addDecorator(withKnobs)
-  .add('Basic Usage', (() => {
-    const data = array('Data',[
-      {sortOrder: 1, population: 2000, label: 'Labrador Retriever'},
-      {sortOrder: 2, population: 8000, label: 'Standard Poodle'},
-      {sortOrder: 3, population: 6000, label: 'French Bulldog'},
-      {sortOrder: 4, population: 3000, label: 'Afghan Hound'},
-      {sortOrder: 5, population: 1000, label: 'Jack Russell Terrier'}
-    ]);
-    const dataKey = text('Data key', 'sortOrder');
-    const dataValue = text('Data values', 'population');
-    const dataKeyLabel = text('Data key labels', 'label');
+const data = array('Data',[
+{sortOrder: 1, population: 2000, label: 'Labrador Retriever'},
+{sortOrder: 2, population: 8000, label: 'Standard Poodle'},
+{sortOrder: 3, population: 6000, label: 'French Bulldog'},
+{sortOrder: 4, population: 3000, label: 'Afghan Hound'},
+{sortOrder: 5, population: 1000, label: 'Jack Russell Terrier'}
+]);
+const dataKey = text('Data key', 'sortOrder');
+const dataValue = text('Data values', 'population');
+const dataKeyLabel = text('Data key labels', 'label');
 
-    return (
+export default () => storiesOf(displayName, module).addDecorator(withKnobs)
+      .add('Basic Usage', (() => {
+        return (
+          <BarChart
+            data={data}
+            dataKey={dataKey}
+            dataValue={dataValue}
+            dataKeyLabel={dataKeyLabel}
+            title={'Dogs and their Money'}
+            subtitle={'As of January 2017'}
+          />
+        );
+      }))
+  .add('No title', (() => (
       <BarChart
         data={data}
         dataKey={dataKey}
         dataValue={dataValue}
         dataKeyLabel={dataKeyLabel}
-        title={'Dogs and their Money'}
-        subtitle={'As of January 2017'}
       />
-    );
-  }))
-  .add('no props', (() => (
-    <BarChart />
   )))
